Revert "Built Core Website with product, collection, and cart pages"
