+++ .eslintrc.js
@@ -1,12 +1,3 @@
-/**
- * @type {import("@types/eslint").Linter.BaseConfig}
- */
 module.exports = {
-  extends: [
-    'plugin:hydrogen/recommended',
-    'plugin:hydrogen/typescript',
-    'prettier',
-  ],
-  plugins: ['prettier'],
-  rules: {'prettier/prettier': ['error', {endOfLine: 'auto'}]},
+  extends: ['plugin:hydrogen/recommended'],
 };
