--- app/root.tsx
@@ -1,91 +1,0 @@
-import {
-  Links,
-  Meta,
-  Outlet,
-  Scripts,
-  ScrollRestoration,
-  useLoaderData,
-} from '@remix-run/react';
-import {type LoaderArgs} from '@shopify/remix-oxygen';
-import tailwind from './styles/tailwind-build.css';
-import favicon from '../public/favicon.svg';
-import {Layout} from './components/Layout';
-import {Seo} from '@shopify/hydrogen';
-import {defer} from '@shopify/remix-oxygen';
-import {CART_QUERY} from '~/queries/cart';
-export const links = () => {
-  return [
-    {rel: 'stylesheet', href: tailwind},
-    {
-      rel: 'preconnect',
-      href: 'https://cdn.shopify.com',
-    },
-    {
-      rel: 'preconnect',
-      href: 'https://shop.app',
-    },
-    {rel: 'icon', type: 'image/svg+xml', href: favicon},
-  ];
-};
-
-export const meta = () => ({
-  charset: 'utf-8',
-  viewport: 'width=device-width,initial-scale=1',
-});
-
-export async function loader({context, request}: LoaderArgs) {
-  const cartId = await context.session.get('cartId');
-
-  return defer({
-    cart: cartId ? getCart(context, cartId) : undefined,
-    layout: await context.storefront.query(LAYOUT_QUERY),
-  });
-}
-
-export default function App() {
-  const data = useLoaderData();
-
-  const {name} = data.layout.shop;
-
-  return (
-    <html lang="en">
-      <head>
-        <Seo />
-        <Meta />
-        <Links />
-      </head>
-      <body>
-        <Layout title={name}>
-          <Outlet />
-        </Layout>
-        <ScrollRestoration />
-        <Scripts />
-      </body>
-    </html>
-  );
-}
-
-const LAYOUT_QUERY = `#graphql
-  query layout {
-    shop {
-      name
-      description
-    }
-  }
-`;
-async function getCart({storefront}: any, cartId: number) {
-  if (!storefront) {
-    throw new Error('missing storefront client in cart query');
-  }
-
-  const {cart} = await storefront.query(CART_QUERY, {
-    variables: {
-      cartId,
-      country: storefront.i18n.country,
-      language: storefront.i18n.language,
-    },
-    cache: storefront.CacheNone(),
-  });
-
-  return cart;
-}
