+++ package.json
@@ -1,30 +1,22 @@
 {
-  "name": "hello-world",
+  "name": "MoshiProject",
   "private": true,
   "sideEffects": false,
   "version": "0.0.0",
   "scripts": {
-    "build": "npm run build:css && shopify hydrogen build --entry ./server",
-    "build:css": "postcss app/styles/tailwind.css -o app/styles/tailwind-build.css --env production",
-    "dev": "npm run build:css && concurrently -g -r npm:dev:css \"shopify hydrogen dev\"",
-    "dev:css": "postcss app/styles/tailwind.css -o app/styles/tailwind-build.css -w",
+    "build": "shopify hydrogen build",
+    "dev": "shopify hydrogen dev",
     "preview": "npm run build && shopify hydrogen preview",
     "lint": "eslint --no-error-on-unmatched-pattern --ext .js,.ts,.jsx,.tsx .",
-    "typecheck": "tsc --noEmit",
     "g": "shopify hydrogen generate"
   },
   "prettier": "@shopify/prettier-config",
   "dependencies": {
-    "@headlessui/react": "^1.7.13",
-    "@heroicons/react": "^2.0.17",
-    "@mantine/carousel": "^6.0.5",
     "@remix-run/react": "1.14.0",
     "@shopify/cli": "3.29.0",
     "@shopify/cli-hydrogen": "^4.0.9",
     "@shopify/hydrogen": "^2023.1.6",
     "@shopify/remix-oxygen": "^1.0.4",
-    "dotenv": "^16.0.3",
-    "embla-carousel-react": "^7.1.0",
     "graphql": "^16.6.0",
     "graphql-tag": "^2.12.6",
     "isbot": "^3.6.6",
@@ -33,22 +25,10 @@
   },
   "devDependencies": {
     "@remix-run/dev": "1.14.0",
-    "@shopify/oxygen-workers-types": "^3.17.2",
     "@shopify/prettier-config": "^1.1.2",
-    "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/forms": "^0.5.3",
-    "@types/eslint": "^8.4.10",
-    "@types/react": "^18.0.20",
-    "@types/react-dom": "^18.0.6",
-    "autoprefixer": "^10.4.14",
-    "concurrently": "^8.0.0",
     "eslint": "^8.20.0",
     "eslint-plugin-hydrogen": "0.12.2",
-    "postcss": "^8.4.21",
-    "postcss-cli": "^10.1.0",
-    "prettier": "^2.8.4",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^4.9.5"
+    "prettier": "^2.8.4"
   },
   "engines": {
     "node": ">=16.13"
