+++ remix.config.js
@@ -3,7 +3,7 @@ module.exports = {
   appDirectory: 'app',
   ignoredRouteFiles: ['**/.*'],
   watchPaths: ['./public'],
-  server: './server.ts',
+  server: './server.js',
   /**
    * The following settings are required to deploy Hydrogen apps to Oxygen:
    */
