--- tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "include": ["./**/*.d.ts", "./**/*.ts", "./**/*.tsx"],
-  "compilerOptions": {
-    "lib": ["DOM", "DOM.Iterable", "ES2022"],
-    "isolatedModules": true,
-    "esModuleInterop": true,
-    "jsx": "react-jsx",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "target": "ES2022",
-    "strict": true,
-    "allowJs": true,
-    "forceConsistentCasingInFileNames": true,
-    "skipLibCheck": true,
-    "baseUrl": ".",
-    "types": ["@shopify/oxygen-workers-types"],
-    "paths": {
-      "~/*": ["app/*"]
-    },
-
-    // Remix takes care of building everything in `./app` with `remix build`.
-    // Wrangler takes care of building everything in `./worker` with `wrangler start` / `wrangler publish`.
-    "noEmit": true
-  }
-}
